i made worker repository and fixed admin rules . . .

+++ LibraryCourseProject/Commands/UserSectionCommands/DeleteCommand.cs
@@ -26,20 +26,19 @@ namespace LibraryCourseProject.Commands.UserSectionCommands
         public void Execute(object parameter)
         {
             var item = UserViewModel.SelectedUser;
-            //  if (item.Id != 1)
-            //  {
-            
-            App.DB.UserRepository.DeleteData(item);
-            UserViewModel.AllUsers = App.DB.UserRepository.GetAllData();
-                //App.Config.Users = new List<User>(UserViewModel.AllUsers);
-                //App.Config.SeriailizeToJson();
-                //App.DB.UserRepository.DeleteData(item);
+            if (item != null && UserViewModel.CurrentUser.Username != "admin")
+            {
+
+                App.DB.UserRepository.DeleteData(item);
+                UserViewModel.AllUsers = App.DB.UserRepository.GetAllData();
                 UserViewModel.SelectedUser = new User();
-           // }
-          //  else
-          //  {
-          //      MessageBox.Show("You can not delete *admin* ");
-//}
+            }
+            else
+            {
+                //data trigger yaz buttonlar sonsun eger admindise
+
+                MessageBox.Show("You can not delete admin");
+            }
 
         }
     }
